// src/app/generated/models/publish-advisory-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface PublishAdvisoryRequest {
  action: 'APPROVED' | 'EDITED' | 'REPLACED';
  diseaseId: string;

  /**
   * The review task version the officer holds
   */
  expectedVersion: number;
  officerNoteBn?: string | null;
  remedyIds: Array<string>;
}